+++ .gitignore
@@ -301,7 +301,7 @@ samples/MicroserviceDemo/microservices/TenantManagementService.Host/Logs/logs.tx
 /npm/packs/bootstrap/package-lock.json
 /templates/app/aspnet-core/src/MyCompanyName.MyProjectName.Web.Host/package-lock.json
 /templates/app/aspnet-core/src/MyCompanyName.MyProjectName.Web/package-lock.json
-/templates/app/aspnet-core/src/MyCompanyName.MyProjectName.IdentityServer/package-lock.json
+/templates/app/aspnet-core/src/MyCompanyName.MyProjectName.AuthServer/package-lock.json
 
 modules/virtual-file-explorer/app/Volo.Abp.VirtualFileExplorer.DemoApp/Logs/
 /templates/app/aspnet-core/src/MyCompanyName.MyProjectName.HttpApi.HostWithIds/package-lock.json
