+++ templates/app/aspnet-core/.gitignore
@@ -254,7 +254,7 @@ paket-files/
 # MyProjectName
 src/MyCompanyName.MyProjectName.Web/Logs/*
 src/MyCompanyName.MyProjectName.Web.Host/Logs/*
-src/MyCompanyName.MyProjectName.IdentityServer/Logs/*
+src/MyCompanyName.MyProjectName.AuthServer/Logs/*
 src/MyCompanyName.MyProjectName.HttpApi.Host/Logs/*
 src/MyCompanyName.MyProjectName.HttpApi.HostWithIds/Logs/*
 src/MyCompanyName.MyProjectName.DbMigrator/Logs/*
