+++ templates/app/aspnet-core/src/MyCompanyName.MyProjectName.AuthServer/MyProjectNameBrandingProvider.cs
