+++ templates/app/aspnet-core/src/MyCompanyName.MyProjectName.AuthServer/Pages/_ViewImports.cshtml
