+++ templates/app/aspnet-core/src/MyCompanyName.MyProjectName.AuthServer/Properties/launchSettings.json
@@ -15,7 +15,7 @@
         "ASPNETCORE_ENVIRONMENT": "Development"
       }
     },
-    "MyCompanyName.MyProjectName.IdentityServer": {
+    "MyCompanyName.MyProjectName.AuthServer": {
       "commandName": "Project",
       "launchBrowser": true,
       "applicationUrl": "https://localhost:44301",
