+++ templates/app/aspnet-core/src/MyCompanyName.MyProjectName.AuthServer/abp.resourcemapping.js
