+++ templates/app/aspnet-core/src/MyCompanyName.MyProjectName.AuthServer/appsettings.json
