+++ templates/app/aspnet-core/src/MyCompanyName.MyProjectName.AuthServer/wwwroot/global-styles.css
