+++ templates/app/aspnet-core/src/MyCompanyName.MyProjectName.Blazor/Menus/MyProjectNameMenuContributor.cs
@@ -52,12 +52,12 @@ public class MyProjectNameMenuContributor : IMenuContributor
     {
         var accountStringLocalizer = context.GetLocalizer<AccountResource>();
 
-        var identityServerUrl = _configuration["AuthServer:Authority"] ?? "";
+        var authServerUrl = _configuration["AuthServer:Authority"] ?? "";
 
         context.Menu.AddItem(new ApplicationMenuItem(
             "Account.Manage",
             accountStringLocalizer["MyAccount"],
-            $"{identityServerUrl.EnsureEndsWith('/')}Account/Manage?returnUrl={_configuration["App:SelfUrl"]}",
+            $"{authServerUrl.EnsureEndsWith('/')}Account/Manage?returnUrl={_configuration["App:SelfUrl"]}",
             icon: "fa fa-cog",
             order: 1000,
             null).RequireAuthenticated());
