+++ templates/module/aspnet-core/.gitignore
@@ -252,7 +252,7 @@ paket-files/
 *.sln.iml
 
 # MyProjectName
-host/MyCompanyName.MyProjectName.IdentityServer/Logs/logs.txt
+host/MyCompanyName.MyProjectName.AuthServer/Logs/logs.txt
 host/MyCompanyName.MyProjectName.HttpApi.Host/Logs/logs.txt
 host/MyCompanyName.MyProjectName.Web.Host/Logs/logs.txt
 host/MyCompanyName.MyProjectName.Web.Unified/Logs/logs.txt
