+++ templates/module/aspnet-core/MyCompanyName.MyProjectName.abpmdl.json
@@ -75,8 +75,8 @@
       "path": "host/MyCompanyName.MyProjectName.Web.Host/MyCompanyName.MyProjectName.Web.Host.abppkg.json",
       "folder": "host"
     },
-    "MyCompanyName.MyProjectName.OpenIddict": {
-      "path": "host/MyCompanyName.MyProjectName.OpenIddict/MyCompanyName.MyProjectName.OpenIddict.abppkg.json",
+    "MyCompanyName.MyProjectName.AuthServer": {
+      "path": "host/MyCompanyName.MyProjectName.AuthServer/MyCompanyName.MyProjectName.AuthServer.abppkg.json",
       "folder": "host"
     },
     "MyCompanyName.MyProjectName.Host.Shared": {
