+++ templates/module/aspnet-core/host/MyCompanyName.MyProjectName.AuthServer/Dockerfile
@@ -5,7 +5,7 @@ EXPOSE 80
 FROM mcr.microsoft.com/dotnet/core/sdk:3.0-buster AS build
 WORKDIR /src
 COPY . .
-WORKDIR /src/templates/service/host/MyCompanyName.MyProjectName.OpenIddict
+WORKDIR /src/templates/service/host/MyCompanyName.MyProjectName.AuthServer
 RUN dotnet restore -nowarn:msb3202,nu1503
 RUN dotnet build --no-restore -c Release -o /app
 
@@ -15,4 +15,4 @@ RUN dotnet publish --no-restore -c Release -o /app
 FROM base AS final
 WORKDIR /app
 COPY --from=publish /app .
-ENTRYPOINT ["dotnet", "MyCompanyName.MyProjectName.OpenIddict.dll"]
+ENTRYPOINT ["dotnet", "MyCompanyName.MyProjectName.AuthServer.dll"]
