+++ templates/module/aspnet-core/host/MyCompanyName.MyProjectName.AuthServer/FodyWeavers.xsd
