+++ templates/module/aspnet-core/host/MyCompanyName.MyProjectName.AuthServer/Pages/Index.cshtml.cs
