+++ templates/module/aspnet-core/host/MyCompanyName.MyProjectName.AuthServer/Pages/_ViewImports.cshtml
