+++ templates/module/aspnet-core/host/MyCompanyName.MyProjectName.AuthServer/Properties/launchSettings.json
@@ -15,7 +15,7 @@
         "ASPNETCORE_ENVIRONMENT": "Development"
       }
     },
-    "OpenIddictHost": {
+    "AuthServerHost": {
       "commandName": "Project",
       "launchBrowser": true,
       "applicationUrl": "https://localhost:44301",
