+++ templates/module/aspnet-core/host/MyCompanyName.MyProjectName.AuthServer/abp.resourcemapping.js
