+++ templates/module/aspnet-core/host/MyCompanyName.MyProjectName.AuthServer/appsettings.json
