+++ templates/module/aspnet-core/host/MyCompanyName.MyProjectName.AuthServer/package.json
@@ -1,6 +1,6 @@
 {
   "version": "1.0.0",
-  "name": "my-app-identityserver",
+  "name": "my-app-authserver",
   "private": true,
   "dependencies": {
     "@abp/aspnetcore.mvc.ui.theme.basic": "^5.3.0-rc.1"
